Indices refactoring and small changes
