+++ xclim/indices/__init__.py
@@ -1,0 +1,7 @@
+import numpy as np
+
+from . import indices
+
+ftomm = np.nan
+
+
