+++ xclim/indices/indices.py
@@ -88,8 +88,6 @@ xr.set_options(enable_cftimeindex=True)  # Set xarray to use cftimeindex
 # Frequencies : YS: year start, QS-DEC: seasons starting in december, MS: month start
 # See http://pandas.pydata.org/pandas-docs/stable/timeseries.html#offset-aliases
 
-ftomm = np.nan
-
 
 # TODO: Define a unit conversion system for temperature [K, C, F] and precipitation [mm h-1, Kg m-2 s-1] metrics
 # TODO: Move utility functions to another file.
@@ -208,48 +206,6 @@ def cold_spell_duration_index(tasmin, tn10, window=6, freq='YS'):
     return below.resample(time=freq).apply(rl.windowed_run_count, window=window, dim='time')
 
 
-@declare_units('days', tas='[temperature]', thresh='[temperature]')
-def cold_spell_days(tas, thresh='-10 degC', window=5, freq='AS-JUL'):
-    r"""Cold spell days
-
-    The number of days that are part of a cold spell, defined as five or more consecutive days with mean daily
-    temperature below a threshold in °C.
-
-    Parameters
-    ----------
-    tas : xarrray.DataArray
-      Mean daily temperature [℃] or [K]
-    thresh : str
-      Threshold temperature below which a cold spell begins [℃] or [K]. Default : '-10 degC'
-    window : int
-      Minimum number of days with temperature below threshold to qualify as a cold spell.
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Cold spell days.
-
-    Notes
-    -----
-    Let :math:`T_i` be the mean daily temperature on day :math:`i`, the number of cold spell days during
-    period :math:`\phi` is given by
-
-    .. math::
-
-       \sum_{i \in \phi} \prod_{j=i}^{i+5} [T_j < thresh]
-
-    where :math:`[P]` is 1 if :math:`P` is true, and 0 if false.
-
-    """
-    t = utils.convert_units_to(thresh, tas)
-    over = tas < t
-    group = over.resample(time=freq)
-
-    return group.apply(rl.windowed_run_count, window=window, dim='time')
-
-
 def cold_and_dry_days(tas, tgin25, pr, wet25, freq='YS'):
     r"""Cold and dry days.
 
@@ -289,63 +245,6 @@ def cold_and_dry_days(tas, tgin25, pr, wet25, freq='YS'):
 
     # c = (c1 * c2) * 1
     # return c.resample(time=freq).sum(dim='time')
-
-
-@declare_units('mm/day', pr='[precipitation]', thresh='[precipitation]')
-def daily_pr_intensity(pr, thresh='1 mm/day', freq='YS'):
-    r"""Average daily precipitation intensity
-
-    Return the average precipitation over wet days.
-
-    Parameters
-    ----------
-    pr : xarray.DataArray
-      Daily precipitation [mm/d or kg/m²/s]
-    thresh : str
-      precipitation value over which a day is considered wet. Default : '1 mm/day'
-    freq : str, optional
-      Resampling frequency defining the periods
-      defined in http://pandas.pydata.org/pandas-docs/stable/timeseries.html#resampling. Default : '1 mm/day'
-
-    Returns
-    -------
-    xarray.DataArray
-      The average precipitation over wet days for each period
-
-    Notes
-    -----
-    Let :math:`\mathbf{p} = p_0, p_1, \ldots, p_n` be the daily precipitation and :math:`thresh` be the precipitation
-    threshold defining wet days. Then the daily precipitation intensity is defined as
-
-    .. math::
-
-       \frac{\sum_{i=0}^n p_i [p_i \leq thresh]}{\sum_{i=0}^n [p_i \leq thresh]}
-
-    where :math:`[P]` is 1 if :math:`P` is true, and 0 if false.
-
-    Examples
-    --------
-    The following would compute for each grid cell of file `pr.day.nc` the average
-    precipitation fallen over days with precipitation >= 5 mm at seasonal
-    frequency, ie DJF, MAM, JJA, SON, DJF, etc.:
-
-    >>> pr = xr.open_dataset('pr.day.nc')
-    >>> daily_int = daily_pr_intensity(pr, thresh='5 mm/day', freq="QS-DEC")
-
-    """
-    t = utils.convert_units_to(thresh, pr, 'hydro')
-
-    # put pr=0 for non wet-days
-    pr_wd = xr.where(pr >= t, pr, 0)
-    pr_wd.attrs['units'] = pr.units
-
-    # sum over wanted period
-    s = pr_wd.resample(time=freq).sum(dim='time', keep_attrs=True)
-    sd = utils.pint_multiply(s, 1 * units.day, 'mm')
-
-    # get number of wetdays over period
-    wd = wetdays(pr, thresh=thresh, freq=freq)
-    return sd / wd
 
 
 @declare_units('days', pr='[precipitation]', thresh='[precipitation]')
@@ -429,86 +328,6 @@ def consecutive_frost_days(tasmin, freq='AS-JUL'):
         frz = units.convert(frz, fu, tu)
     group = (tasmin < frz).resample(time=freq)
     return group.apply(rl.longest_run, dim='time')
-
-
-@declare_units('days', pr='[precipitation]', thresh='[precipitation]')
-def maximum_consecutive_wet_days(pr, thresh='1 mm/day', freq='YS'):
-    r"""Consecutive wet days.
-
-    Returns the maximum number of consecutive wet days.
-
-    Parameters
-    ---------
-    pr : xarray.DataArray
-      Mean daily precipitation flux [Kg m-2 s-1] or [mm]
-    thresh : str
-      Threshold precipitation on which to base evaluation [Kg m-2 s-1] or [mm]. Default : '1 mm/day'
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      The maximum number of consecutive wet days.
-
-    Notes
-    -----
-    Let :math:`\mathbf{x}=x_0, x_1, \ldots, x_n` be a daily precipitation series and
-    :math:`\mathbf{s}` be the sorted vector of indices :math:`i` where :math:`[p_i > thresh] \neq [p_{i+1} >
-    thresh]`, that is, the days when the precipitation crosses the *wet day* threshold.
-    Then the maximum number of consecutive wet days is given by
-
-    .. math::
-
-
-       \max(\mathbf{d}) \quad \mathrm{where} \quad d_j = (s_j - s_{j-1}) [x_{s_j} > 0\celsius]
-
-    where :math:`[P]` is 1 if :math:`P` is true, and 0 if false. Note that this formula does not handle sequences at
-    the start and end of the series, but the numerical algorithm does.
-    """
-    thresh = utils.convert_units_to(thresh, pr, 'hydro')
-
-    group = (pr > thresh).resample(time=freq)
-    return group.apply(rl.longest_run, dim='time')
-
-
-@declare_units('C days', tas='[temperature]', thresh='[temperature]')
-def cooling_degree_days(tas, thresh='18 degC', freq='YS'):
-    r"""Cooling degree days
-
-    Sum of degree days above the temperature threshold at which spaces are cooled.
-
-    Parameters
-    ----------
-    tas : xarray.DataArray
-      Mean daily temperature [℃] or [K]
-    thresh : str
-      Temperature threshold above which air is cooled. Default : '18 degC'
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Cooling degree days
-
-    Notes
-    -----
-    Let :math:`x_i` be the daily mean temperature at day :math:`i`. Then the cooling degree days above
-    temperature threshold :math:`thresh` over period :math:`\phi` is given by:
-
-    .. math::
-
-        \sum_{i \in \phi} (x_{i}-{thresh} [x_i > thresh]
-
-    where :math:`[P]` is 1 if :math:`P` is true, and 0 if false.
-    """
-    thresh = utils.convert_units_to(thresh, tas)
-
-    return tas.pipe(lambda x: x - thresh) \
-        .clip(min=0) \
-        .resample(time=freq) \
-        .sum(dim='time')
 
 
 @declare_units('days', tasmax='[temperature]', tasmin='[temperature]')
@@ -661,48 +480,6 @@ def extreme_temperature_range(tasmax, tasmin, freq='YS'):
     return out
 
 
-@declare_units('', tas='[temperature]', thresh='[temperature]')
-def freshet_start(tas, thresh='0 degC', window=5, freq='YS'):
-    r"""First day consistently exceeding threshold temperature.
-
-    Returns first day of period where a temperature threshold is exceeded
-    over a given number of days.
-
-    Parameters
-    ----------
-    tas : xarray.DataArray
-      Mean daily temperature [℃] or [K]
-    thresh : str
-      Threshold temperature on which to base evaluation [℃] or [K]. Default '0 degC'
-    window : int
-      Minimum number of days with temperature above threshold needed for evaluation
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    float
-      Day of the year when temperature exceeds threshold over a given number of days for the first time. If there are
-      no such day, return np.nan.
-
-    Notes
-    -----
-    Let :math:`x_i` be the daily mean temperature at day of the year :math:`i` for values of :math:`i` going from 1
-    to 365 or 366. The start date of the freshet is given by the smallest index :math:`i` for which
-
-    .. math::
-
-       \prod_{j=i}^{i+w} [x_j > thresh]
-
-    is true, where :math:`w` is the number of days the temperature threshold should be exceeded,  and :math:`[P]` is
-    1 if :math:`P` is true, and 0 if false.
-    """
-    thresh = utils.convert_units_to(thresh, tas)
-    over = (tas > thresh)
-    group = over.resample(time=freq)
-    return group.apply(rl.first_run_ufunc, window=window, index='dayofyear')
-
-
 @declare_units('days', tasmin='[temperature]')
 def frost_days(tasmin, freq='YS'):
     r"""Frost days index
@@ -737,129 +514,6 @@ def frost_days(tasmin, freq='YS'):
         frz = units.convert(frz, fu, tu)
     f = (tasmin < frz) * 1
     return f.resample(time=freq).sum(dim='time')
-
-
-@declare_units('C days', tas='[temperature]', thresh='[temperature]')
-def growing_degree_days(tas, thresh='4.0 degC', freq='YS'):
-    r"""Growing degree-days over threshold temperature value [℃].
-
-    The sum of degree-days over the threshold temperature.
-
-    Parameters
-    ---------
-    tas : xarray.DataArray
-      Mean daily temperature [℃] or [K]
-    thresh : str
-      Threshold temperature on which to base evaluation [℃] or [K]. Default: '4.0 degC'.
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      The sum of growing degree-days above 4℃
-
-    Notes
-    -----
-    Let :math:`TG_{ij}` be the daily mean temperature at day :math:`i` of period :math:`j`. Then the
-    growing degree days are:
-
-    .. math::
-
-        GD4_j = \sum_{i=1}^I (TG_{ij}-{4} | TG_{ij} > {4}℃)
-    """
-    thresh = utils.convert_units_to(thresh, tas)
-    return tas.pipe(lambda x: x - thresh) \
-        .clip(min=0) \
-        .resample(time=freq) \
-        .sum(dim='time')
-
-
-@declare_units('days', tas='[temperature]', thresh='[temperature]')
-def growing_season_length(tas, thresh='5.0 degC', window=6, freq='YS'):
-    r"""Growing season length.
-
-    The number of days between the first occurrence of at least
-    six consecutive days with mean daily temperature over 5℃ and
-    the first occurrence of at least six consecutive days with
-    mean daily temperature below 5℃ after July 1st in the northern
-    hemisphere and January 1st in the southern hemisphere.
-
-    Parameters
-    ---------
-    tas : xarray.DataArray
-      Mean daily temperature [℃] or [K]
-    thresh : str
-      Threshold temperature on which to base evaluation [℃] or [K]. Default: '5.0 degC'.
-    window : int
-      Minimum number of days with temperature above threshold to mark the beginning and end of growing season.
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Growing season length.
-
-    Notes
-    -----
-    Let :math:`TG_{ij}` be the mean temperature at day :math:`i` of period :math:`j`. Then counted is
-    the number of days between the first occurrence of at least 6 consecutive days with:
-
-    .. math::
-
-        TG_{ij} > 5 ℃
-
-    and the first occurrence after 1 July of at least 6 consecutive days with:
-
-    .. math::
-
-        TG_{ij} < 5 ℃
-    """
-
-    # i = xr.DataArray(np.arange(tas.time.size), dims='time')
-    # ind = xr.broadcast(i, tas)[0]
-    #
-    # c = ((tas > thresh) * 1).rolling(time=window).sum()
-    # i1 = ind.where(c == window).resample(time=freq).min(dim='time')
-    #
-    # # Resample sets the time to T00:00.
-    # i11 = i1.reindex_like(c, method='ffill')
-    #
-    # # TODO: Adjust for southern hemisphere
-    #
-    # #i2 = ind.where(c == 0).where(tas.time.dt.month >= 7)
-    # # add check to make sure indice of end of growing season is after growing season start
-    # i2 = ind.where((c==0) & (ind > i11)).where(tas.time.dt.month >= 7)
-    #
-    # d = i2 - i11
-    #
-    # # take min value (first occurence after july)
-    # gsl = d.resample(time=freq).min(dim='time')
-    #
-    # # turn nan into 0
-    # gsl = xr.where(np.isnan(gsl), 0, gsl)
-
-    # compute growth season length on resampled data
-    thresh = utils.convert_units_to(thresh, tas)
-
-    c = ((tas > thresh) * 1).rolling(time=window).sum().chunk(tas.chunks)
-
-    def compute_gsl(c):
-        nt = c.time.size
-        i = xr.DataArray(np.arange(nt), dims='time').chunk({'time': 1})
-        ind = xr.broadcast(i, c)[0].chunk(c.chunks)
-        i1 = ind.where(c == window).min(dim='time')
-        i1 = xr.where(np.isnan(i1), nt, i1)
-        i11 = i1.reindex_like(c, method='ffill')
-        i2 = ind.where((c == 0) & (ind > i11)).where(c.time.dt.month >= 7)
-        i2 = xr.where(np.isnan(i2), nt, i2)
-        d = (i2 - i1).min(dim='time')
-        return d
-
-    gsl = c.resample(time=freq).apply(compute_gsl)
-
-    return gsl
 
 
 @declare_units('', tasmin='[temperature]', tasmax='[temperature]', thresh_tasmin='[temperature]',
@@ -918,35 +572,6 @@ def heat_wave_frequency(tasmin, tasmax, thresh_tasmin='22.0 degC', thresh_tasmax
     cond = (tasmin > thresh_tasmin) & (tasmax > thresh_tasmax)
     group = cond.resample(time=freq)
     return group.apply(rl.windowed_run_events, window=window, dim='time')
-
-
-@declare_units('days', tasmax='[temperature]', thresh='[temperature]')
-def heat_wave_index(tasmax, thresh='25.0 degC', window=5, freq='YS'):
-    r"""Heat wave index.
-
-    Number of days that are part of a heatwave, defined as five or more consecutive days over 25℃.
-
-    Parameters
-    ----------
-    tasmax : xarrray.DataArray
-      Maximum daily temperature [℃] or [K]
-    thresh : str
-      Threshold temperature on which to designate a heatwave [℃] or [K]. Default: '25.0 degC'.
-    window : int
-      Minimum number of days with temperature above threshold to qualify as a heatwave.
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    DataArray
-      Heat wave index.
-    """
-    thresh = utils.convert_units_to(thresh, tasmax)
-    over = tasmax > thresh
-    group = over.resample(time=freq)
-
-    return group.apply(rl.windowed_run_count, window=window, dim='time')
 
 
 @declare_units('days', tasmin='[temperature]', tasmax='[temperature]', thresh_tasmin='[temperature]',
@@ -1008,43 +633,6 @@ def heat_wave_max_length(tasmin, tasmax, thresh_tasmin='22.0 degC', thresh_tasma
     group = cond.resample(time=freq)
     max_l = group.apply(rl.longest_run, dim='time')
     return max_l.where(max_l >= window, 0)
-
-
-@declare_units('C days', tas='[temperature]', thresh='[temperature]')
-def heating_degree_days(tas, thresh='17.0 degC', freq='YS'):
-    r"""Heating degree days
-
-    Sum of degree days below the temperature threshold at which spaces are heated.
-
-    Parameters
-    ----------
-    tas : xarray.DataArray
-      Mean daily temperature [℃] or [K]
-    thresh : str
-      Threshold temperature on which to base evaluation [℃] or [K]. Default: '17.0 degC'.
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Heating degree days index.
-
-    Notes
-    -----
-    Let :math:`TG_{ij}` be the daily mean temperature at day :math:`i` of period :math:`j`. Then the
-    heating degree days are:
-
-    .. math::
-
-        HD17_j = \sum_{i=1}^{I} (17℃ - TG_{ij})
-    """
-    thresh = utils.convert_units_to(thresh, tas)
-
-    return tas.pipe(lambda x: thresh - x) \
-        .clip(0) \
-        .resample(time=freq) \
-        .sum(dim='time')
 
 
 @declare_units('days', tasmax='[temperature]')
@@ -1135,74 +723,6 @@ def liquid_precip_ratio(pr, prsn=None, tas=None, freq='QS-DEC'):
     rain = tot - prsn.resample(time=freq).sum(dim='time')
     ratio = rain / tot
     return ratio
-
-
-@declare_units('days', tasmin='[temperature]', thresh='[temperature]')
-def tn_days_below(tasmin, thresh='-10.0 degC', freq='YS'):
-    r"""Number of days with tmin below a threshold in
-
-    Number of days where daily minimum temperature is below a threshold.
-
-    Parameters
-    ----------
-    tasmin : xarray.DataArray
-      Minimum daily temperature [℃] or [K]
-    thresh : str
-      Threshold temperature on which to base evaluation [℃] or [K] . Default: '-10 degC'.
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Number of days Tmin < threshold.
-
-    Notes
-    -----
-    Let :math:`TN_{ij}` be the daily minimum temperature at day :math:`i` of period :math:`j`. Then
-    counted is the number of days where:
-
-    .. math::
-
-        TX_{ij} < Threshold [℃]
-    """
-    thresh = utils.convert_units_to(thresh, tasmin)
-    f1 = utils.threshold_count(tasmin, '<', thresh, freq)
-    return f1
-
-
-@declare_units('days', tasmax='[temperature]', thresh='[temperature]')
-def tx_days_above(tasmax, thresh='25.0 degC', freq='YS'):
-    r"""Number of summer days
-
-    Number of days where daily maximum temperature exceed a threshold.
-
-    Parameters
-    ----------
-    tasmax : xarray.DataArray
-      Maximum daily temperature [℃] or [K]
-    thresh : str
-      Threshold temperature on which to base evaluation [℃] or [K]. Default: '25 degC'.
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Number of summer days.
-
-    Notes
-    -----
-    Let :math:`TX_{ij}` be the daily maximum temperature at day :math:`i` of period :math:`j`. Then
-    counted is the number of days where:
-
-    .. math::
-
-        TX_{ij} > Threshold [℃]
-    """
-    thresh = utils.convert_units_to(thresh, tasmax)
-    f = (tasmax > (thresh)) * 1
-    return f.resample(time=freq).sum(dim='time')
 
 
 @declare_units('mm', pr='[precipitation]')
@@ -1478,109 +998,6 @@ def tg10p(tas, t10, freq='YS'):
     return below.resample(time=freq).sum(dim='time')
 
 
-@declare_units('[temperature]', tas='[temperature]')
-def tg_max(tas, freq='YS'):
-    r"""Highest mean temperature.
-
-    The maximum of daily mean temperature.
-
-    Parameters
-    ----------
-    tas : xarray.DataArray
-      Mean daily temperature [℃] or [K]
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Maximum of daily minimum temperature.
-
-    Notes
-    -----
-    Let :math:`TN_{ij}` be the mean temperature at day :math:`i` of period :math:`j`. Then the maximum
-    daily mean temperature for period :math:`j` is:
-
-    .. math::
-
-        TNx_j = max(TN_{ij})
-    """
-
-    return tas.resample(time=freq).max(dim='time', keep_attrs=True)
-
-
-@declare_units('[temperature]', tas='[temperature]')
-def tg_mean(tas, freq='YS'):
-    r"""Mean of daily average temperature.
-
-    Resample the original daily mean temperature series by taking the mean over each period.
-
-    Parameters
-    ----------
-    tas : xarray.DataArray
-      Mean daily temperature [℃] or [K]
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      The mean daily temperature at the given time frequency
-
-    Notes
-    -----
-    Let :math:`TN_i` be the mean daily temperature of day :math:`i`, then for a period :math:`p` starting at
-    day :math:`a` and finishing on day :math:`b`:
-
-    .. math::
-
-       TG_p = \frac{\sum_{i=a}^{b} TN_i}{b - a + 1}
-
-
-    Examples
-    --------
-    The following would compute for each grid cell of file `tas.day.nc` the mean temperature
-    at the seasonal frequency, ie DJF, MAM, JJA, SON, DJF, etc.:
-
-    >>> t = xr.open_dataset('tas.day.nc')
-    >>> tg = tm_mean(t, freq="QS-DEC")
-    """
-
-    arr = tas.resample(time=freq) if freq else tas
-    return arr.mean(dim='time', keep_attrs=True)
-
-
-@declare_units('[temperature]', tas='[temperature]')
-def tg_min(tas, freq='YS'):
-    r"""Lowest mean temperature
-
-    Minimum of daily mean temperature.
-
-    Parameters
-    ----------
-    tas : xarray.DataArray
-      Mean daily temperature [℃] or [K]
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Minimum of daily minimum temperature.
-
-    Notes
-    -----
-    Let :math:`TG_{ij}` be the mean temperature at day :math:`i` of period :math:`j`. Then the minimum
-    daily mean temperature for period :math:`j` is:
-
-    .. math::
-
-        TGn_j = min(TG_{ij})
-    """
-
-    return tas.resample(time=freq).min(dim='time', keep_attrs=True)
-
-
 # TODO: Improve description
 @declare_units('days', tasmin='[temperature]', t90='[temperature]')
 def tn90p(tasmin, t90, freq='YS'):
@@ -1675,100 +1092,6 @@ def tn10p(tasmin, t10, freq='YS'):
     below = (tasmin < thresh)
 
     return below.resample(time=freq).sum(dim='time')
-
-
-@declare_units('[temperature]', tasmin='[temperature]')
-def tn_max(tasmin, freq='YS'):
-    r"""Highest minimum temperature.
-
-    The maximum of daily minimum temperature.
-
-    Parameters
-    ----------
-    tasmin : xarray.DataArray
-      Minimum daily temperature [℃] or [K]
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Maximum of daily minimum temperature.
-
-    Notes
-    -----
-    Let :math:`TN_{ij}` be the minimum temperature at day :math:`i` of period :math:`j`. Then the maximum
-    daily minimum temperature for period :math:`j` is:
-
-    .. math::
-
-        TNx_j = max(TN_{ij})
-    """
-
-    return tasmin.resample(time=freq).max(dim='time', keep_attrs=True)
-
-
-@declare_units('[temperature]', tasmin='[temperature]')
-def tn_mean(tasmin, freq='YS'):
-    r"""Mean minimum temperature.
-
-    Mean of daily minimum temperature.
-
-    Parameters
-    ----------
-    tasmin : xarray.DataArray
-      Minimum daily temperature [℃] or [K]
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Mean of daily minimum temperature.
-
-    Notes
-    -----
-    Let :math:`TN_{ij}` be the minimum temperature at day :math:`i` of period :math:`j`. Then mean
-    values in period :math:`j` are given by:
-
-    .. math::
-
-        TN_{ij} = \frac{ \sum_{i=1}^{I} TN_{ij} }{I}
-    """
-
-    arr = tasmin.resample(time=freq) if freq else tasmin
-    return arr.mean(dim='time', keep_attrs=True)
-
-
-@declare_units('[temperature]', tasmin='[temperature]')
-def tn_min(tasmin, freq='YS'):
-    r"""Lowest minimum temperature
-
-    Minimum of daily minimum temperature.
-
-    Parameters
-    ----------
-    tasmin : xarray.DataArray
-      Minimum daily temperature [℃] or [K]
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Minimum of daily minimum temperature.
-
-    Notes
-    -----
-    Let :math:`TN_{ij}` be the minimum temperature at day :math:`i` of period :math:`j`. Then the minimum
-    daily minimum temperature for period :math:`j` is:
-
-    .. math::
-
-        TNn_j = min(TN_{ij})
-    """
-
-    return tasmin.resample(time=freq).min(dim='time', keep_attrs=True)
 
 
 @declare_units('days', tasmin='[temperature]', thresh='[temperature]')
@@ -1904,134 +1227,6 @@ def tx10p(tasmax, t10, freq='YS'):
     return below.resample(time=freq).sum(dim='time')
 
 
-@declare_units('[temperature]', tasmax='[temperature]')
-def tx_max(tasmax, freq='YS'):
-    r"""Highest max temperature
-
-    The maximum value of daily maximum temperature.
-
-    Parameters
-    ----------
-    tasmax : xarray.DataArray
-      Maximum daily temperature [℃] or [K]
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Maximum value of daily maximum temperature.
-
-    Notes
-    -----
-    Let :math:`TX_{ij}` be the maximum temperature at day :math:`i` of period :math:`j`. Then the maximum
-    daily maximum temperature for period :math:`j` is:
-
-    .. math::
-
-        TXx_j = max(TX_{ij})
-    """
-
-    return tasmax.resample(time=freq).max(dim='time', keep_attrs=True)
-
-
-@declare_units('[temperature]', tasmax='[temperature]')
-def tx_mean(tasmax, freq='YS'):
-    r"""Mean max temperature
-
-    The mean of daily maximum temperature.
-
-    Parameters
-    ----------
-    tasmax : xarray.DataArray
-      Maximum daily temperature [℃] or [K]
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Mean of daily maximum temperature.
-
-    Notes
-    -----
-    Let :math:`TX_{ij}` be the maximum temperature at day :math:`i` of period :math:`j`. Then mean
-    values in period :math:`j` are given by:
-
-    .. math::
-
-        TX_{ij} = \frac{ \sum_{i=1}^{I} TX_{ij} }{I}
-    """
-
-    arr = tasmax.resample(time=freq) if freq else tasmax
-    return arr.mean(dim='time', keep_attrs=True)
-
-
-@declare_units('[temperature]', tasmax='[temperature]')
-def tx_min(tasmax, freq='YS'):
-    r"""Lowest max temperature
-
-    The minimum of daily maximum temperature.
-
-    Parameters
-    ----------
-    tasmax : xarray.DataArray
-      Maximum daily temperature [℃] or [K]
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Minimum of daily maximum temperature.
-
-    Notes
-    -----
-    Let :math:`TX_{ij}` be the maximum temperature at day :math:`i` of period :math:`j`. Then the minimum
-    daily maximum temperature for period :math:`j` is:
-
-    .. math::
-
-        TXn_j = min(TX_{ij})
-    """
-
-    return tasmax.resample(time=freq).min(dim='time', keep_attrs=True)
-
-
-@declare_units('days', tasmax='[temperature]', thresh='[temperature]')
-def warm_day_frequency(tasmax, thresh='30 degC', freq='YS'):
-    r"""Frequency of extreme warm days
-
-    Return the number of days with tasmax > thresh per period
-
-    Parameters
-    ----------
-    tasmax : xarray.DataArray
-      Mean daily temperature [℃] or [K]
-    thresh : str
-      Threshold temperature on which to base evaluation [℃] or [K]. Default : '30 degC'
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      Number of days exceeding threshold.
-
-    Notes:
-    Let :math:`TX_{ij}` be the daily maximum temperature at day :math:`i` of period :math:`j`. Then
-    counted is the number of days where:
-
-    .. math::
-
-        TN_{ij} > Threshold [℃]
-
-    """
-    thresh = utils.convert_units_to(thresh, tasmax)
-    events = (tasmax > thresh) * 1
-    return events.resample(time=freq).sum(dim='time')
-
-
 @declare_units('days', tasmin='[temperature]', tasmax='[temperature]', thresh_tasmin='[temperature]',
                thresh_tasmax='[temperature]')
 def tx_tn_days_above(tasmin, tasmax, thresh_tasmin='22 degC',
@@ -2084,31 +1279,6 @@ def tx_tn_days_above(tasmin, tasmax, thresh_tasmin='22 degC',
     return events.resample(time=freq).sum(dim='time')
 
 
-@declare_units('days', tasmin='[temperature]', thresh='[temperature]')
-def warm_night_frequency(tasmin, thresh='22 degC', freq='YS'):
-    r"""Frequency of extreme warm nights
-
-    Return the number of days with tasmin > thresh per period
-
-    Parameters
-    ----------
-    tasmin : xarray.DataArray
-      Minimum daily temperature [℃] or [K]
-    thresh : str
-      Threshold temperature on which to base evaluation [℃] or [K]. Default : '22 degC'
-    freq : str, optional
-      Resampling frequency
-
-    Returns
-    -------
-    xarray.DataArray
-      The number of days with tasmin > thresh per period
-    """
-    thresh = utils.convert_units_to(thresh, tasmin, )
-    events = (tasmin > thresh) * 1
-    return events.resample(time=freq).sum(dim='time')
-
-
 @declare_units('days', tasmax='[temperature]', tx90='[temperature]')
 def warm_spell_duration_index(tasmax, tx90, window=6, freq='YS'):
     r"""Warm spell duration index
@@ -2157,41 +1327,6 @@ def warm_spell_duration_index(tasmax, tx90, window=6, freq='YS'):
     above = (tasmax > thresh)
 
     return above.resample(time=freq).apply(rl.windowed_run_count, window=window, dim='time')
-
-
-@declare_units('days', pr='[precipitation]', thresh='[precipitation]')
-def wetdays(pr, thresh='1.0 mm/day', freq='YS'):
-    r"""Wet days
-
-    Return the total number of days during period with precipitation over threshold.
-
-    Parameters
-    ----------
-    pr : xarray.DataArray
-      Daily precipitation [mm]
-    thresh : str
-      Precipitation value over which a day is considered wet. Default: '1 mm/day'.
-    freq : str, optional
-      Resampling frequency defining the periods
-      defined in http://pandas.pydata.org/pandas-docs/stable/timeseries.html#resampling.
-
-    Returns
-    -------
-    xarray.DataArray
-      The number of wet days for each period [day]
-
-    Examples
-    --------
-    The following would compute for each grid cell of file `pr.day.nc` the number days
-    with precipitation over 5 mm at the seasonal frequency, ie DJF, MAM, JJA, SON, DJF, etc.:
-
-    >>> pr = xr.open_dataset('pr.day.nc')
-    >>> wd = wetdays(pr, pr_min = 5., freq="QS-DEC")
-    """
-    thresh = utils.convert_units_to(thresh, pr, 'hydro')
-
-    wd = (pr >= thresh) * 1
-    return wd.resample(time=freq).sum(dim='time')
 
 
 @declare_units('', pr='[precipitation]', prsn='[precipitation]', tas='[temperature]')
