+++ xclim/subset.py
